Ballot SC21: Log Integrity Controls
Revise the Network and Certificate System Security Requirements section
3.e. to allow for automated monitoring, edit section 3.f., and add
section 3.g. to establish a response time for automated alerts.

Purpose of Ballot:

The Network and Certificate System Security Requirements committee is
proposing this ballot to revise the current requirements to better allow
for automation in the monitoring of systems. The goal of this ballot is
to remove manual efforts that can be less effective and more
resource-intensive than automated monitoring and alerting.

This ballot also adds specific requirements in terms of the timeliness
for addressing alerting from automated monitoring and alerting to ensure
the implementation of automation does not increase the length of time
that a potential issue could go without being detected.

Proposed by Ben Wilson of DigiCert and endorsed by Trevoli Ponds-White
of Amazon and Fotis Loukos of SSL.com.