Ballot SC20: System Configuration Management
Purpose of Ballot:

Section 1(h) of the Network and Certification Systems Security
Requirements provides that CAs shall:

    Review configurations of Issuing Systems, Certificate Management
    Systems, Security Support Systems, and Front-End / Internal-Support
    Systems on at least a weekly basis to determine whether any changes
    violated the CA’s security policies;

In relation to this requirement the WebTrust/PKI Assurance Task Force found and recommended that:

    Systems are too complex to perform a meaningful human review of
    configuration changes. Software monitoring tools play a large part in
    achieving this requirement, but there is a required human element to
    inspect the monitoring software as well as consider changes in physical
    security. Consider better detailing the goals of this criterion.

This ballot seeks to implement the recommendation of the WebTrust/PKI
Assurance Task Force by expressly allowing CAs to rely on continuous
monitoring instead of human reviews.

Proposed by Ben Wilson of DigiCert and endorsed by Tobias Josefowitz of
OPERA and Dustin Hollenback of Microsoft.