Ballot SCXX: Logging and Log Retention
Modifies both the BR and NSR.

Proposed by X, endorsed by Y and Z.

Effective as of the date Ballot SCXX passes.

Background
==========

Both the Baseline Requirements for the Issuance and Management of
Publicly-Trusted Certificates (BR) and the Network and Certificate
System Security Requirements (NSR) stipulate that CAs must collect and
retain log files concerning a number of events. These events are
separated into three categories, CA key lifecycle management events,
Subscriber Certificate lifecycle management events, and security events.

While the purpose for the retention requirements for the CA key
lifecycle management events and Subscriber Certificate lifecycle
management events is well understood, the purpose for the same retention
period for security events is less clear. Security event logs decrease
in relevancy and value as time passes from the time of the event. The
requirements for what is included in security events in ambiguous and
can generate a significant amount of log data. The current requirements
require CAs to maintain a significant amount of data that CAs consider
to be of low value, and has a significant amount of compliance overhead.

Purpose
=======

The proposed changes seek to clarify the relationship between audit
logging obligations under Network and Certification System Security
Requirements and Baseline Requirements and to reduce the retention
period for non-issuance related log data. The proposed change also
provides clarification by specifically cross-referencing the Baseline
Requirements.

Certificate Lifecycle Events
----------------------------

Common record retention periods were considered when drafting this
ballot. For transaction history the United States Internal Revenue
Service and medical transaction history require retention periods of
five to seven years.  Some state requirements can require up to ten
years of transaction log history. European countries requirements vary,
but in most cases are between five and seven years for business records.
Based on these common requirements, seven years of CA and subscriber
certificate lifecycle management and CA key lifecycle management events
is reasonable compared to various record retention requirements.   

Security Events
---------------

Most data retention standards do not specify retention requirements for
the underlying systems that process transactions. In drafting this
ballot it was considered the primary use of these system logs is in the
event of an incident and an investigation is required. The IBM and
Ponemon Institute's 2019 Cost of a Data Breach Report found on average
it took 206 days to identify a data breach had occurred. CA environments
are expected to be more secure and include above average monitoring. To
provide enough history for incident investigation two years of
supporting system logs will provide more than enough time to detect and
investigate the incident.

Risk vs Benefits
================

The benefit of this ballot is to reduce data retention requirements for
those log elements most CAs consider having little value. As an example
firewall and router activity logs of significant size and storage
requirements. These logs serve a benefit when investigating a potential
security event, however these logs lose value with the passage of time.
Logs contain firewall traffic that is several years old provides little
value in the investigation of a potential incident. The log size for
many CAs is measured in terabytes, each year and the overhead of storing
these logs and monitoring for compliance is significant. The benefit for
reducing retention is considered high.

The primary risk of accepting this ballot is a CA experiences a security
event, does not identify the event for more than two years and does not
have the required logs needed to fully investigate the impact of the
event. A CA environment is considered a high security environment and is
required to have robust security monitoring and alerting. There is a
reduced likelihood a CA would have a security event go undetected for
more than two years. In the event a security event was undetected for a
year, the reliability of the logs is also decreased. Overall the risks
associated with this ballot are low.