Ballot SCXX: System account management
Sesction 2(j) is modified to allow for review or continuous monitoring,
system accounts identified by either means now need to be deactivated
within 24 hours. Previously, deactivation of such accounts was to be
done as part of the review with no (further) directions.

Purpose of Ballot:

Section 2(j) of the Network and Certificate System Security Requirements
("NSRs") require:

    Each CA or Delegated Third Party SHALL: [...] Review all system
    accounts at least every three (3) months and deactivate any accounts
    that are no longer necessary for operations;

Section 2 and 2(j) are (more) specific provisions about system accounts,
which means that they are not to be treated as "configuration" in
regards to section 1(h) of the NSRs. Ballot SC 20 aims to "provide CAs
with greater certainty on what automated system configuration monitoring
satisfies the goal of Section 1(h)". This ballot aims to provide a
compatible change to Section 2(j), based on the conviction that
implementing and monitoring defined system states systematically and
with automated systems should generally be encouraged and results in
more well-maintained and thus more secure systems.